+++ src/main/java/com/gxdxx/instagram/domain/chatroom/api/ChatRoomController.java
@@ -2,7 +2,7 @@ package com.gxdxx.instagram.domain.chatroom.api;
 
 import com.gxdxx.instagram.domain.chatroom.dto.request.ChatRoomListRequest;
 import com.gxdxx.instagram.domain.chatroom.dto.response.ChatRoomResponse;
-import com.gxdxx.instagram.global.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.chatroom.application.ChatRoomQueryService;
 import io.swagger.v3.oas.annotations.Operation;
