+++ src/main/java/com/gxdxx/instagram/domain/comment/api/CommentController.java
@@ -4,8 +4,8 @@ import com.gxdxx.instagram.domain.comment.dto.request.CommentRegisterRequest;
 import com.gxdxx.instagram.domain.comment.dto.request.CommentUpdateRequest;
 import com.gxdxx.instagram.domain.comment.dto.response.CommentRegisterResponse;
 import com.gxdxx.instagram.domain.comment.dto.response.CommentUpdateResponse;
-import com.gxdxx.instagram.global.dto.response.ErrorResponse;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.comment.application.CommentCreateService;
 import com.gxdxx.instagram.domain.comment.application.CommentDeleteService;
