+++ src/main/java/com/gxdxx/instagram/domain/comment/application/CommentDeleteService.java
@@ -1,6 +1,6 @@
 package com.gxdxx.instagram.domain.comment.application;
 
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.comment.domain.Comment;
 import com.gxdxx.instagram.domain.comment.exception.CommentNotFoundException;
 import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
