+++ src/main/java/com/gxdxx/instagram/domain/message/api/MessageController.java
@@ -2,9 +2,9 @@ package com.gxdxx.instagram.domain.message.api;
 
 import com.gxdxx.instagram.domain.message.dto.request.MessageListRequest;
 import com.gxdxx.instagram.domain.message.dto.request.MessageSendRequest;
-import com.gxdxx.instagram.global.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
 import com.gxdxx.instagram.domain.message.dto.response.MessageResponse;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.message.application.MessageCreateService;
 import com.gxdxx.instagram.domain.message.application.MessageQueryService;
