+++ src/main/java/com/gxdxx/instagram/domain/post/api/PostController.java
@@ -6,8 +6,8 @@ import com.gxdxx.instagram.domain.post.dto.request.PostUpdateRequest;
 import com.gxdxx.instagram.domain.post.dto.response.FeedResponse;
 import com.gxdxx.instagram.domain.post.dto.response.PostRegisterResponse;
 import com.gxdxx.instagram.domain.post.dto.response.PostUpdateResponse;
-import com.gxdxx.instagram.global.dto.response.ErrorResponse;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.post.application.PostCreateService;
 import com.gxdxx.instagram.domain.post.application.PostDeleteService;
