+++ src/main/java/com/gxdxx/instagram/domain/post/application/PostUpdateService.java
@@ -6,7 +6,7 @@ import com.gxdxx.instagram.domain.post.domain.Post;
 import com.gxdxx.instagram.domain.post.exception.PostNotFoundException;
 import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
 import com.gxdxx.instagram.domain.post.dao.PostRepository;
-import com.gxdxx.instagram.config.s3.S3Uploader;
+import com.gxdxx.instagram.global.config.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
