+++ src/main/java/com/gxdxx/instagram/domain/reply/api/ReplyController.java
@@ -4,8 +4,8 @@ import com.gxdxx.instagram.domain.reply.dto.request.ReplyRegisterRequest;
 import com.gxdxx.instagram.domain.reply.dto.request.ReplyUpdateRequest;
 import com.gxdxx.instagram.domain.reply.dto.response.ReplyRegisterResponse;
 import com.gxdxx.instagram.domain.reply.dto.response.ReplyUpdateResponse;
-import com.gxdxx.instagram.global.dto.response.ErrorResponse;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.reply.application.ReplyCreateService;
 import com.gxdxx.instagram.domain.reply.application.ReplyDeleteService;
