+++ src/main/java/com/gxdxx/instagram/domain/user/application/UserCreateService.java
@@ -5,7 +5,7 @@ import com.gxdxx.instagram.domain.user.dto.response.UserSignUpResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
-import com.gxdxx.instagram.config.s3.S3Uploader;
+import com.gxdxx.instagram.global.config.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
