+++ src/main/java/com/gxdxx/instagram/domain/user/application/UserDeleteService.java
@@ -1,6 +1,6 @@
 package com.gxdxx.instagram.domain.user.application;
 
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.PasswordNotMatchException;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
