+++ src/main/java/com/gxdxx/instagram/global/auth/AccessTokenCreateService.java
@@ -1,7 +1,7 @@
 package com.gxdxx.instagram.global.auth;
 
-import com.gxdxx.instagram.config.jwt.TokenProvider;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.config.jwt.TokenProvider;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
