+++ src/main/java/com/gxdxx/instagram/global/config/audit/AuditorAwareImpl.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.config;
+package com.gxdxx.instagram.global.config.audit;
 
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.security.core.Authentication;
