+++ src/main/java/com/gxdxx/instagram/global/config/jwt/TokenProvider.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.config.jwt;
+package com.gxdxx.instagram.global.config.jwt;
 
 import com.gxdxx.instagram.domain.user.domain.User;
 import io.jsonwebtoken.Claims;
