+++ src/main/java/com/gxdxx/instagram/global/config/s3/S3Config.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.config.s3;
+package com.gxdxx.instagram.global.config.s3;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
