+++ src/test/java/com/gxdxx/instagram/service/follow/FollowCreateServiceTest.java
@@ -2,7 +2,7 @@ package com.gxdxx.instagram.service.follow;
 
 import com.gxdxx.instagram.domain.follow.application.FollowCreateService;
 import com.gxdxx.instagram.domain.follow.dto.request.FollowCreateRequest;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.follow.exception.FollowAlreadyExistsException;
