+++ src/test/java/com/gxdxx/instagram/service/follow/FollowDeleteServiceTest.java
@@ -1,7 +1,7 @@
 package com.gxdxx.instagram.service.follow;
 
 import com.gxdxx.instagram.domain.follow.application.FollowDeleteService;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.follow.exception.FollowNotFountException;
