+++ src/test/java/com/gxdxx/instagram/service/post/PostCreateServiceTest.java
@@ -8,7 +8,7 @@ import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.post.dao.PostRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
-import com.gxdxx.instagram.config.s3.S3Uploader;
+import com.gxdxx.instagram.global.config.s3.S3Uploader;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
