+++ src/test/java/com/gxdxx/instagram/service/user/UserCreateServiceTest.java
@@ -6,7 +6,7 @@ import com.gxdxx.instagram.domain.user.dto.response.UserSignUpResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
-import com.gxdxx.instagram.config.s3.S3Uploader;
+import com.gxdxx.instagram.global.config.s3.S3Uploader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
