+++ src/test/java/com/gxdxx/instagram/service/user/UserDeleteServiceTest.java
@@ -1,7 +1,7 @@
 package com.gxdxx.instagram.service.user;
 
 import com.gxdxx.instagram.domain.user.application.UserDeleteService;
-import com.gxdxx.instagram.global.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.domain.user.exception.PasswordNotMatchException;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
